updated a frontend file structure
