+++ client/Navigation.jsx
@@ -4,10 +4,9 @@ import { Link } from 'react-router-dom';
 const Navigation = () => {
   return (
     <nav>
+      {/* add link to each page */}
       <Link to="/">Home</Link>
       <Link to="/dog">Dog</Link>
-      <Link to="/signup">Signup</Link>
-      <Link to="/login">Login</Link>
     </nav>
   );
 };
